Database issue fixed

+++ estore/src/main/java/com/api/swed/estore/dao/ProductDAO.java
@@ -6,7 +6,7 @@ import com.api.swed.estore.models.Product;
 
 public interface ProductDAO {
 
-	Product get(int productId);
+	Product get(long productId);
 	List<Product> list();	
 	Product add(Product product);
 	boolean update(Product product);
